+++ tests/compile-and-dump/Promote/Prelude.hs
@@ -4,6 +4,7 @@ import Data.Promotion.TH
 import Data.Promotion.Prelude
 import Data.Promotion.Prelude.List
 import Data.Proxy
+import GHC.TypeLits
 
 lengthTest1a :: Proxy (Length '[True, True, True, True])
 lengthTest1a = Proxy